Fetched the data for Retrieve Question by ID from the mock API
