+++ app/src/main/java/com/revature/popquiz/model/api/services/QuizApiService.kt
@@ -2,9 +2,9 @@ package com.revature.popquiz.model.api.services
 
 import com.revature.popquiz.model.api.services.popquiz.RequestPopQuiz
 import com.revature.popquiz.model.api.services.popquiz.ResponsePopQuiz
+import com.revature.popquiz.model.api.services.questions.ResponseRetrieveQuestionByID
 import com.revature.popquiz.model.api.services.quiz.RequestAllQuizzes
 import com.revature.popquiz.model.api.services.quiz.ResponseAllQuizzes
-import com.revature.popquiz.model.api.services.quiz.ResponseRetrieveQuestionByID
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
