+++ app/src/main/java/com/revature/popquiz/model/api/services/quiz/ResponseAllQuizzes.kt
@@ -15,7 +15,6 @@ data class ResponseAllQuizzes (
     @SerializedName("data")
     var quizList:List<QuizResponse>
 
-
 )
 data class QuizResponse(
 
@@ -27,34 +26,5 @@ data class QuizResponse(
     var sCategoryName:String
 )
 
-data class ResponseRetrieveQuestionByID(
 
-    @SerializedName("qIdStatusCode")
-    var qIdStatusCode: Int,
-    @SerializedName("qIdDescription")
-    var qIdDescription: String,
-    @SerializedName("qIdData")
-    var retrieveQuestionByIDList: List<QuestionsResponse>
-)
 
-data class QuestionsResponse(
-
-    @SerializedName("questionId")
-    var questionId: Int,
-    @SerializedName("questionTitle")
-    var questionTitle: String,
-    @SerializedName("questionLevelId")
-    var questionLevelId: Int,
-    var questionAnswers: List<QuestionAnswersResponse>
-)
-
-data class QuestionAnswersResponse(
-    @SerializedName("questionAnswerId")
-    var questionAnswerId: Int,
-    @SerializedName("questionAnswer")
-    var questionAnswer: String,
-    @SerializedName("questionAnswerOrder")
-    var questionAnswerOrder: Int,
-    @SerializedName("questionAnswerCorrect")
-    var questionAnswerCorrect: Boolean
-)
